requirements.txt
scrapy==1.8.0
selenium==2.53.6
pyvirtualdisplay
image
tldextract
twisted==16.6
requests
python-dateutil
scrapy-random-useragent
scrapy-fake-useragent
scrapy_crawlera
bs4
paramiko
fuzzywuzzy
psycopg2-binary
SQLAlchemy==1.3.1
sqlalchemy_utils==0.33.9
dnspython
pymongo
pymongo[srv]
boto3
botocore
pathlib
pandas
xlwt
xlrd
openpyxl
mysqlclient==1.3.4
loginform
pyOpenSSL
future
lxml
scraperapi-sdk